app/requirements.txt
flask==3.0.0
requests==2.31.0
octodns==1.8.0
pyyaml==6.0.1
